Using the Key instead of Caption
Because the Enumeration keys cannot contain dashes ('-'), and because Charset.forName() must receive the correct name (that includes dashes), I used replace() to switch underscores to dashes in the key and pass the result to forName().
getCaption() could have worked, but only for projects having en-US as a default language.